Refactoring meta generation
